mcp-textpseudomizer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-textpseudomizer"
version = "0.1.0"
description = "MCP server for automatic text pseudonymization using Named Entity Recognition"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.1.0",
    "torch>=2.0.0",
    "langdetect>=1.0.9",
    "dateparser>=1.1.8",
    "phonenumbers>=8.13.0",
    "email-validator>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "uvicorn>=0.22.0",
]

[project.optional-dependencies]
# Full ML dependencies (may require compilation on some platforms)
ml = [
    "flair>=0.12.0",
    "spacy>=3.6.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]
# Minimal dependencies for local development without heavy ML dependencies
dev-minimal = [
    "fastmcp>=0.1.0",
    "langdetect>=1.0.9",
    "dateparser>=1.1.8", 
    "phonenumbers>=8.13.0",
    "email-validator>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]
